+++ server/src-lib/Hasura/Backends/MSSQL/Instances/Schema.hs
@@ -323,6 +323,35 @@ msComparisonExps = P.memoize 'comparisonExps \columnType -> do
         (Just "does the column NOT match the given pattern")
         (ANLIKE    . mkParameter <$> typedParser)
       ]
+
+    -- Ops for Geometry/Geography types
+    , guard (isScalarColumnWhere (`elem` MSSQL.geoTypes) columnType) *>
+      [ P.fieldOptional $$(G.litName "_st_contains")
+        (Just "does the column contain the given value")
+        (ASTContains   . mkParameter <$> typedParser)
+      , P.fieldOptional $$(G.litName "_st_equals")
+        (Just "is the column equal to given value (directionality is ignored)")
+        (ASTEquals     . mkParameter <$> typedParser)
+      , P.fieldOptional $$(G.litName "_st_intersects")
+        (Just "does the column spatially intersect the given value")
+        (ASTIntersects     . mkParameter <$> typedParser)
+      , P.fieldOptional $$(G.litName "_st_overlaps")
+        (Just "does the column 'spatially overlap' (intersect but not completely contain) the given value")
+        (ASTOverlaps     . mkParameter <$> typedParser)
+      , P.fieldOptional $$(G.litName "_st_within")
+        (Just "is the column contained in the given value")
+        (ASTWithin     . mkParameter <$> typedParser)
+      ]
+
+    -- Ops for Geometry types
+    , guard (isScalarColumnWhere (MSSQL.GeometryType ==) columnType) *>
+      [ P.fieldOptional $$(G.litName "_st_crosses")
+        (Just "does the column cross the given geometry value")
+        (ASTCrosses   . mkParameter <$> typedParser)
+      , P.fieldOptional $$(G.litName "_st_touches")
+        (Just "does the column have at least one point in common with the given geometry value")
+        (ASTTouches     . mkParameter <$> typedParser)
+      ]
     ]
   where
     mkListLiteral :: [ColumnValue 'MSSQL] -> UnpreparedValue 'MSSQL